+++ lib/card_dealer.rb
@@ -9,5 +9,5 @@ end
 require_relative "card_dealer/version"
 require_relative "card_dealer/card"
 require_relative "card_dealer/deck"
-require_relative "card_dealer/deck_builder"
+require_relative "card_dealer/build_deck"
 require_relative "card_dealer/binary_deck"